Update patch set 1
Patch Set 1:

Wow, calm down. If I was rejecting this, there would be a -1 attached. 

I'm ok with merging this in the interim, but I would prefer Ethan's comments be addressed as a long-term goal. 

I do have competing priorities though, so just relax and quit taking non-action as a personal slight.

Patch-set: 1