Update patch set 1
Patch Set 1:

You guys are right, I shouldn't have an opinion. I should just keep my mouth shut and wait for Steve and Koush to tell me what to say and do.........OH WAIT, I'm not a mindless idiot! hahaha

I'm not a CM employee or subject to some CM code of conduct or  whatever the fuck so why don't you guys not tell me what I can and can't say and you continue on "not speaking" because that's what Steve and Koush wants you to do....OK?

Patch-set: 1