Update patch set 5
Patch Set 5:

Is the work necessary to continue supporting the Galaxy Nexus unreasonable as compared to a device like the Galaxy S2 or RAZR HD/M (msm8969) devices?

Are there are extra conditions around the OMAP platform (as compared to other legacy platforms), or is the primary roadblock simply a lack of help?

Patch-set: 5