Update patch set 5
Patch Set 5:

OFC you can choose different ways on your projects. What mads me here is the repeated "we dont give a fuck" coming from CM heads regarding what is done on XDA, like if XDA (and it userbase, and developers) didn't helped a lot made CM into what it is now. You should, at least for the sake of respect, matter to what goes on there. In this case going on here, Daniel just used XDA (and others) as a simple example that people care about platform so it should be of CM's interest to. Nobody said that you must copy everything that the world is doing. You refuted this with nothing, with the plain "i don't care" like if his argument was not valid.

Look, I don't even have the Gnex. But some kind of attitude just disgusts. And Ive been with CM for long time now...

Patch-set: 5