Update patch set 6
Patch Set 6:

@Justin, the issue here is two-fold:

1) Lack of personal time (I did maguro, Brint did toro, and Michael did toroplus). Ziyan has stepped up for maguro, but we are still down folks who can invest the necessary time into the other two (hence their removal). 

2) The OMAP platform is end of life from TI (the chipset manufacturer). This means a complete lack of meaningful updates from the silicon vendor themselves. Projects do exist to try and extend the lifecycle of the chip (ie forking OMAPZoom's last drop and adapting it), but this is a serious time sink - which relates back to item #1

The mxm8xxx are all well supported (and Qualcomm's chips are fairly easy to work with). The Exynos (Samsung) chip may not be as well documented, but there are enough resources (and frankly, people who own the devices since Samsungs Android footprint is so damn high) that the time commitment factor becomes trivial. Coupled with Exynos based projects like HardKernel, the issues are nowhere near hat seen on OMAP. 

OMAP4 as seen on Moto devices has stargo backing it, so I am not making a blanket 'remove all OMAP' action at this time.

Patch-set: 6