Update patch set 6
Patch Set 6:

@Pedro "I don't care about XDA/Omni/Insert ROM here"

You seem to have missed the 'here' as in 'in this context of this patch'. Offering XDA or an external project as a counter to my "no one is actively caring for this platform" (which was in the context of CM maintainers) is a no-op as far as justifications go. 

"Daniel just used XDA (and others) as a simple example that people care about platform so it should be of CM's interest to."

That line of reasoning is really off the mark. People care about lots of platforms (AllWinner, Mediatek, Rockchip, etc) - doesn't mean we will. 

Frankly, reading into my comment as a 'CM disrespects XDA' is baseless. Hell, I'll be in Manchester at the end of the month for XDADevcon. I don't really have much else to say on that topic as that doesn't belong here at all.

Patch-set: 6