Update patch set 2
Patch Set 2:

This commit only enables _nightlies_. It does not _fix_ anything. I don't want to turn on a nightly build for a device which we do not have someone actually paying attention to. There were already problems with the unified JF/HLTE builds being broken for certain variants for months on nightly builds.

Patch-set: 2