Update patch set 2
Patch Set 2:

If you enable nightlies for kltedv then you can fix our NFC issue. It's not cool that we don't get a good build. Battery life is crap on klte for 900i due to NFC. We all have tested it and will continue to test for kltedv once you enable it.

Patch-set: 2