Update patch set 1
Patch Set 1: Code-Review+1

OK so after further testing and coming from a clean install this bug does not appear. Must be something left over that is not getting reset when doing a dirty flash. I just confirmed this myself and with a few users.

Just flashed D852 using strictly CM sources with no mods and all is well.

Patch-set: 1