Update patch set 1
Patch Set 1: Code-Review-1

If there's not an actual maintainer who owns and tests the device, we shouldn't ship nightlies for it. I realize its very similar to the other gsm jf variants, but the bs with fragmented firmware is way to much to just assume its working at any given time.

Patch-set: 1