Update patch set 1
Patch Set 1:

The questions about it were in posts on the CM 12 for jflteatt on xda-developers. Many of them are modifying the updater script to install on their phone, others are using a custom recovery that has the product.device set to "jflte" instead of "jfltecan". One has asked about unifying jflteatt and jfltecan.

I made a post yesterday telling those that have jfltecan that if they want official nightlies with no hacked install method, someone needs to step up to maintain it.

Patch-set: 1