Update patch set 2
Patch Set 2: Code-Review-1

That's a completely different story. You'll need to get in contact with someone higher up and get the branches made, if and only if, d803 compiles and is functioning on cm-12.0. A bringup will need to be done for it, if it hasn't been already.

Patch-set: 2
Label: Code-Review=-1