Update patch set 1
Patch Set 1:

Lack of maintainers, I know finnq is working on d855, but we don't have anyone looking at the others.

We could possibly ship d850, since its more-or-less identical to d852, but d851 and ls990 are actually quite a bit different (broadcom instead of wcnss, and in the case of ls990: eww, sprint).

End result is I'm not going to be responsible for a variant that I don't own anymore, and I wouldn't ask anyone else to be.

Patch-set: 1