Update patch set 1
Patch Set 1:

A2DP is most defintiely a deal breaker. I must have forgot to test it or something when I shipped jfltevzw. Rolling a build with Arne's patches now...

Regardless of all that, I was under the impression you weren't going to maintain this device anymore. We aren't going to ship it without a maintainer.

Patch-set: 1