Update patch set 4
Patch Set 4:

> FWIW A2DP was working on dlx (apq8064) and I just noticed it
 > doesn't have DEVICE_OUT_PROXY in its audio_policy.conf, wasn't
 > using the header, and has an a2dp section in its audio_policy.conf.
 > 
 > https://github.com/CyanogenMod/android_device_htc_dlx/blob/cm-13.0/configs/audio_policy.conf#L56-L65
 > 
 > I'm not going to pretend to know shit about audio, but I'll
 > probably do some more looking when I get home.

I added the a2dp portion to jf audio_policy and that fixed a2dp. Raising volume on the phone was a little funky, but it played fine. I was only able to test it momentarily, because I had to try it on a friend's car. Listened to a song and it played fine.

Patch-set: 4