Update patch set 2
Patch Set 2: Code-Review+1

After countless unofficial nightlies and hours spent on this, maguro is now stable enough to get cm-12.1 builds. The only "major" issue left is the 8-12 pixel wide (depending on the resolution, 720p or 1080p) pink line on the right side sometimes appearing on recorded videos, which we can hopefully fix later (it's pretty complicated and not too annoying, I don't want to delay because of that issue).

Patch-set: 2
Label: Code-Review=+1