Update patch set 1
Patch Set 1:

I keep saying I will look at this or that, but in reality I just don't have time to keep supporting klte. If someone else wants to be officially responsible for the klte family, or knows someone who wants to be officially responsible for it, please get in contact with me.

Patch-set: 1