Update patch set 3
Patch Set 3: Code-Review-2

When a build for an unknown device gets added to get.cm, Jenkins creates a placeholder entry in the database containing only the model name. In the template, this happens: 

      {% if device.image_url  %}
        <img src={{device.image_url}} />
      {% else %}
        <img src="/static/img/no-image.jpg">
      {% endif %}

So even if no devices use it, the codebase still does for devices that haven't been updated on the server yet with accurate information. I'm more than happy to accept a patch for the README so people won't realize this exists, though.

Patch-set: 3