Update patch set 5
Patch Set 5:

The light sensor not working is probably not a total dealbreaker, but it also spams the logs and I think that's causing eating up CPU cycles which causes occasional camera problems in addition to the main one I fixed. I'm working on an open source sensors HAL which has a working light sensor (and a broken everything else lol).

Like I said, not a total dealbreaker, but otherwise the new sensors HAL should be ready in a day or so. I can only dev about 2 nights a week at the moment (tonight is one of those nights)

Patch-set: 5