Update patch set 1
Patch Set 1: Code-Review-1

As maintainers and the ones who did the majority of the bring up for this device, Brett or I will submit this when we feel it is ready. SElinux in enforcing is not stable even with your commit, it needs more work.

Patch-set: 1
Reviewer: Scott Brissenden <2608@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1