Update patch set 1
Patch Set 1:

hmm, i wonder if all these denials are due to older firmware? i hypothesize this is true. none of these look related to rotation through.

the timerirq denial is probably fixed by adding this to file_contexts:
/dev/timerirq  u:object_r:sensors_device:s0

for the rild denials, it would be nice to know the sysfs nodes it's attempting to write, instead of granting write to all sysfs nodes.

Patch-set: 1