Update patch set 1
Patch Set 1: Code-Review-1

Requires:

http://review.cyanogenmod.org/147845
http://review.cyanogenmod.org/#/c/153360/
http://review.cyanogenmod.org/#/c/153942/

After those changes, the only major outstanding issues are flip cases / "smartview cases" and the fact that we can't enable the proximity check on wakeup.

Note that I've punted on fixing hdr support and just removed it to work around it.

Pushing this up now but I would like to merge the selinux commit and let xda test it for a couple of days before talking about adding this to the nightlies.

Patch-set: 1