Update patch set 1
Patch Set 1:

> Did you try the mic when making a call in non-speaker phone
 > (handset mode) as well? I found that mic was barley audible when I
 > tried it.
 > 
 > Also I believe that screen mirroring refers to casting to
 > chromecast or a miracast receiver.

I did not have issues with the regular handset mic. I merged the speakerphone mic change.

Casting to a Nexus Player works fine in my testing, but casting to WiFi Display appears broken. Best I can tell, it's related to p2p. Our init service and p2p overlay look correct, so if p2p is to blame, it is probably an issue with our ancient-and-heavily-htc-hacked-up bcmdhd_4335_mcc driver or the netmgrd blob.

Patch-set: 1