Update patch set 1
Patch Set 1:

> > > As far as I know the Google Camera (3.2.042/5) should also work
 > > on
 > > > Marshmallow (API 23)
 > >
 > > Yes, you're right; 2.5 was designed for lollipop. I actually do
 > not
 > > see Google Camera available in the Play store when I search. I
 > can
 > > only find version 2.5 because it is in my list of previously
 > > installed apps.
 > 
 > Yes, Google hasn't added the app to the play store. (I downloaded
 > the app outside the playstore).
 > 
 > This could be the app: https://android.googlesource.com/platform/packages/apps/Camera2/+/android-6.0.1_r68

Oh, I had assumed Camera2 was verified working before the switch to Snap, but this doesn't appear to be the case: switch to snap on june 10 and hexedit cam blob on june 15. So, yes, I suppose we should verify Camera2 does still work. I do not consider Google Camera 3.2 to be high priority since it is not available in the Play store and must be manually installed.

Patch-set: 1