Update patch set 1
Patch Set 1:

Wifi region code option is broken on M7. This problem was there on 12.1 branch as well.
Region code gets reset after some uptime and after a reboot. And without it, I can't connect to 5ghz and 2.4ghz N networks. The device is only capable of connecting to 2.4ghz in G mode (OFDM). 
As far as I know, the driver sets the region automatically  based on sim card inserted, and the carrier I'm using (Velcom, Belarus) is not whitelisted to connect to N/AC. Because of this when I'm using stock sence rom I can't connect to N no matter what I try. CM allows to change the region code, but the option doesn't get saved.

Patch-set: 1