Update patch set 1
Patch Set 1:

> > Sadly, the region code gone reset again. http://i.imgur.com/UOB66GQ.png
 > >
 > > I suppose, this commit http://review.cyanogenmod.org/#/c/57588/
 > > could help, but it's probably too aggressive to just remove the
 > > simcard check completely. Maybe it's possible to add Belarus to
 > the
 > > whitelist so when the phone sees MCC code=257, it allows for N/AC
 > > networks to get connected to?
 > 
 > I suppose I would need to see a log, as MccTable has a legitimate
 > entry for Belarus and the static reg database enabled in the kernel
 > also contains Belarus.

What logs do I have to provide, exactly?
https://1drv.ms/u/s!AlEgaKT2cJbtj5MriYf8OexSZGMQfA
Judging by this, the phone sets the region code to BY and this blocks it from connecting to N/AC. At the bottom of the file I think it's the part where I set the region to Europe (GB). And then wifi works fine.

Patch-set: 1