Update patch set 2
Patch Set 2:

Is it possible to integrate the fix for bad white balance calibration for devices with OV4688? OV4688 is the replacement main camera unit for M7, the same sensor is found in M8. The problem is that M7 has bad calibration settings for OV.
There is a method of replacing awb_cal for the one from M8 after every boot http://forum.xda-developers.com/showthread.php?t=2744757 but may be it's possible to integrate proper calibration from M8 directly into the cm's M7 kernel?

Patch-set: 2