Update patch set 1
Patch Set 1:

> Is there anything else needed here before this can land?

It needs a maintainer. Someone who can build/test/debug/fix issues that come up with it. In this case its not as big of an undertaking as some other devices since its very similar to other g3 variants, but we still need someone with a d852 who is willing to address bugs on it.

Patch-set: 1
Label: Code-Review=0