Update patch set 1
Patch Set 1:

The audio volume is a lot lower in comparison to what it was in older rom releases - according to posts on xda even in google's stock rom. I upgraded my flounder to harryyoud's last release (04/22/17). I will See if my netflix problem still persists on that release.

Patch-set: 1