Update patch set 2
Patch Set 2:

So I had a chat with @razorloves, and I can reproduce the noise when playing Youtube videos in Gello browser. However, in Chrome, it doesn't happen. Did not check Jelly yet.
In any case, I don't consider this a blocker. It's a bug which can be fixed any time in the future, but nothing that prevents me from using this version...

Patch-set: 2