Update patch set 2
Patch Set 2:

I really don't see why this is still blocked. Like I already said: This issue does not happen inside the Youtube app nor in Chrome, it only happens in Gello. Someone may want to check if it also happens in Jelly or not, but actually I don't think it matters much...
So, what are we waiting for?

Patch-set: 2