Update patch set 2
Patch Set 2:

Apparently there is some disagreement on whether or not prebuilt audio is a blocker.  I personally do not think that it is.

It is, however, a pain for you and long term support so you would really want to get rid of it as soon as possible.

I looked at audio.primary.msm8916.so and where it uses realcall and it seems like it could be as simple as calling Voip_setRealCall(x) where x is 1 or 0 for on and off.

But again, I don't think that this should be a blocker for shipping the device.

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0