Update patch set 1
Patch Set 1:

I've run the maintainers script several times, and it passes. For good measure I will also make an unofficial build and test that one to check that everything works as intended, as it should be because nothing has changed for the device since it has been dropped.

Patch-set: 1