Update patch set 1
Patch Set 1:

The last time we went through this, listing on the *licensor's* site was the definitive answer. And lacking that, the COA that presents least risk of a cease-and-decist is the most prudent action.

Only when devices are listed on the aptx website should they ship the blobs.

Patch-set: 1