Update patch set 1
Patch Set 1:

aptx.com is the nearest thing to a list of devices that have paid the licensing fee. A vendor shipping blobs may or may not mean they did so. Wait for listing on aptx.com or contact mother Qualcomm and ask if a device you are curious about is actually licensed.

Patch-set: 1