Update patch set 1
Patch Set 1: Code-Review+1

Functionality mirrors 14.1 (except L1, which isn't mandatory), and we were able to support the in-remote mic this cycle (though its bound to break sooner or later, as its finicky, it'll be removed when/if it breaks).

Dependent on vendor change merge. Will ping here when its in.

Patch-set: 1
Label: Code-Review=+1