Update patch set 1
Patch Set 1:

sw or hw because hw is not needed because it is also not supported in stock! the only issue i have is that the atomic shim is sometimes not loaded properly. after fixing this there shouldn't be anything else needed for nightlies

Patch-set: 1
Reviewer: Oskar Kratochvil <14088@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0