Update patch set 1
Patch Set 1:

My main issue with this approach is its verbosity. IMO it would be better if we had some mapping of "device name => android_device_$VENDOR_$TREE": since the targetpath is always going to be "device/$VENDOR/$TREE", there's no need to explicitly specify it.
If the device is not found in this mapping then fall back to the current target_path/repository logic in roomservice as-is (which would work for 99% of devices).

I think the possibility of shared codenames between vendors is definitely something that should be fixed, but probably separately to this series.

Patch-set: 1
Reviewer: Simon Shields <11604@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>