Update patch set 3
Patch Set 3: Code-Review-1

I am not particularly against this, but I think we should highlight what is at stake with the decision of merging this.

AFAIK, crackling has currently no device maintainer for 15.1: the bulk of the work was done by Bestas (kudos for that), but he doesn't have a device to actually test things. Up to 14.1 crackling was "maintained" by me, and, as some might have noticed, I haven't actually do a good job at it, as I don't have the time needed to dedicate to this device.

So, I ask:
a) do we really add 15.1 support to a device without maintainer?
b) is Bestas (or anyone else) interested in becoming the maintainer for this device, despite not having one?
c) if the answer to (b) is yes, then the question becomes: do we want to add support to a device (and release builds for it) that is maintained blindly?

To reiterate, my -1 isn't quite an answer to these questions, but in hopes that some discussion happens about these issues before this is decided to move on (either merged or abandoned).

Also, since this only applies to crackling, I have nothing against splitting this patch in two, and have kipper promoted to 15.1.

Patch-set: 3
Reviewer: Marcos Marado <19730@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1