Update patch set 2
Patch Set 2:

While not having an active maintainer is bad, it's not actually the biggest problem here. According to the below, our official installation instructions result in an inoperable device: 
https://www.reddit.com/r/LineageOS/comments/9z6idc/the_wrong_installing_instructions_on_the_official/ 

If performing a `fastboot erase userdata` instead of a TWRP format is the correct procedure, then the wiki needs to be updated to describe that.

Fact is, the incorrect wiki instructions should have prevented whyred from shipping. The device will be pulled from the roster, if only temporarily, if instructions are not corrected five days from now.

Patch-set: 2
CC: Kevin Haggerty <15173@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>