Update patch set 2
Patch Set 2:

While yes i have announced that I am not going to maintain this device anymore, but the problem with wiki is not from my side. The instructions for "formatting data" are part of the template which i cannot change and following that step leads to bootloop on this device since this uses fde. The commit that added the aforementioned steps was added after this device was added to wiki. Since i did not get enough time to look at it again, I never knew the instructions have become incorrect.

Patch-set: 2
CC: Vasishath Kaushal <14151@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>