Update patch set 2
Patch Set 2:

> Fact is, the incorrect wiki instructions should have prevented whyred from shipping. The device will be pulled from the roster, if only temporarily, if instructions are not corrected five days from now.

I hoped that such hints would be in the wiki, as soon as they are known. That would have prevented much hassle installing LineageOS on my whyred. (But as I now know how complicated it is to change the wiki, I guess I now know why it wasn't updated. What a shame.)

But I finally succeeded installing LineageOS, and I've created updated install-instructions and a pull request: https://github.com/LineageOS/lineage_wiki/pull/105
As soon as this is merged, the install-instructions are working again.

Patch-set: 2
CC: Gerrit User 20788 <20788@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>