Update patch set 2
Patch Set 2:

> Yes, stop policing gerrit change messages.

This commit was not coming up with this search https://review.lineageos.org/q/guacamoleb but is now because of my comment.

What is the + side of not mentioning the device's codename in the commit message?

Many users search the Gerrit for clues about what is happening with their devices. Not including any keywords is...

Patch-set: 2