Update patch set 2
Patch Set 2:

> > Gerrit is not a user facing tool - it's a developer tool.
> 
> *user facing -> user tool

Quote from Wiki - Requesting LineageOS for your device https://wiki.lineageos.org/device_requests.html :

"Perhaps the only somewhat reliable indicator of whether a device may be officially supported is when a developer’s hard work in bringing-up a device is forked over to github.com/LineageOS."

And if you search the LineageOS subreddit with Gerrit, it returns hundreds of results.

Hence the importance of including the device codename in the commit description so it comes up in users' Gerrit searches.

Patch-set: 2